# Reduce the impact of Vec::reserve calls that do not cause any allocation