@the8472 2021-03-21
> This also produces perf increases on the order of 3-5% in a few other microbenchmarks that I'm tracking.

`alloc` has a whole array of benchmarks for vec too. You can run them via `./x.py bench library/alloc --test-args vec::` and diff the results with [cargo benchcmp](https://github.com/BurntSushi/cargo-benchcmp).